refactor: change ID types from int to Long for Course, Room, and Teacher entities

+++ src/main/java/com/college/Course.java
@@ -10,7 +10,7 @@ public class Course {
     @Id
     @Column(name = "course_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int courseId;
+    private Long courseId;
 
     @Column(name = "course_name")
     private String courseName;
@@ -24,11 +24,11 @@ public class Course {
     private List<ClassSchedule> classSchedules;
 
     // Getters and Setters
-    public int getCourseId() {
+    public Long getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(int courseId) {
+    public void setCourseId(Long courseId) {
         this.courseId = courseId;
     }
